Redesign
-Use jwt gem to generate tokens
-Use jwt to manage token expiration
-Use conventional headers
-change User#username to #email

+++ app/controllers/toke/sessions_controller.rb
@@ -1,26 +1,51 @@
 module Toke
   class SessionsController < ApplicationController
-
     before_action :toke, only: :destroy
 
     def create
-      user = User.find_by username: params[:session][:username]
-      if user && user.authenticate(params[:session][:password])
-        user.toke
-        render json: user, status: 201
+      @user = authenticate_with_http_basic do |email, password|
+        user = User.find_by email: email
+        if user && user.authenticate(password)
+          user.token.destroy if user.token
+          user.token = Token.create expires_at: 1.year.from_now
+          user.token.generate_key!
+          user.token.save
+          user
+        end
+      end
+      if @user
+        response.headers['Authorization'] = "Token #{@user.token.key}"
+        render json: @user, status: :created
       else
-        head :unauthorized
+        render json: { Unauthorized: 'Invalid email or password' }, status: :unauthorized
+      end
+    end
+
+    def update
+      token = authenticate_with_http_token do |jwt, options|
+        Token.decode(jwt)[0]
+      end
+      user = token.user if token
+      if user
+        response.headers['Authorization'] = "Token #{token.key}"
+        render json: user
+      else
+        render json: { Unauthorized: 'Invalid session' }, status: :unauthorized
       end
     end
 
     def destroy
-      token = Token.find_by(id: params[:id])
-      if token
-        token.destroy
-        head :no_content
-      else
-        head :not_found
-      end
+      @user.token.destroy
+      head :no_content
+    end
+
+    private
+
+    def payload
+      {
+        user_id: @user.id,
+        exp: 1.year.from_now.to_i
+      }
     end
   end
 end
